requirements.txt
Django==4.1.5
crispy-bootstrap5==0.7
django-crispy-forms==1.14.0
Pillow==9.4.0

